requirements.txt
# Core Dependencies - Strands Agents with Gemini
strands-agents[gemini]==1.15.0
strands-agents-tools==0.2.14
python-dotenv==1.0.1

# FastAPI and Web Server
fastapi==0.115.6
uvicorn==0.38.0
pydantic==2.12.4
pydantic-settings==2.12.0


# Jupyter Notebook Support
jupyter==1.1.1
ipykernel==6.29.5
ipywidgets==8.1.3
notebook==7.2.2

# Async and Networking
httpx==0.28.1
aiofiles==24.1.0

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1  # for testing FastAPI

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2

# Monitoring and Logging
prometheus-client==0.20.0
python-json-logger==2.0.7

# Optional: Rate Limiting
slowapi==0.1.9

# Optional: Database Support (if needed)
# sqlalchemy==2.0.36
# alembic==1.14.0
# psycopg2-binary==2.9.10  # for PostgreSQL

# Development Tools
pre-commit==3.8.0
watchfiles==1.0.3
